requirements.txt
fastapi
uvicorn==0.19.0
python-dotenv
google-api-python-client
requests
yt-dlp